requirements.txt
lark-oapi>=1.0
openai>=1.0
fastapi>=0.110
uvicorn>=0.29
httpx>=0.27
pyyaml>=6.0
